use a common type for web3-provider-engine - also considering using typeRoots instead of paths? web client shouldn't need to know about a type that wc-provider uses internally, unsure how to bundle though
