+++ packages/cardpay-cli/package.json
@@ -31,7 +31,6 @@
     "@types/bn.js": "^5.1.0",
     "@types/lodash": "^4.14.168",
     "@types/semver": "^7.3.9",
-    "@types/web3-provider-engine": "^14.0.0",
     "@walletconnect/web3-provider": "^1.6.0",
     "jsonapi-typescript": "^0.1.3",
     "lodash": "^4.17.21",
