+++ packages/cardpay-cli/tsconfig.json
@@ -14,6 +14,9 @@
     "noUnusedParameters": true,
     "skipLibCheck": true,
     "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true
+    "noImplicitReturns": true,
+    "paths": {
+      "web3-provider-engine": ["../../types/web3-provider-engine"]
+    }
   },
 }
