+++ packages/cardpay-sdk/package.json
@@ -20,7 +20,6 @@
     "@types/lodash": "^4.14.168",
     "@types/node-fetch": "^2.5.10",
     "@types/url-parse": "^1.4.4",
-    "@types/web3-provider-engine": "^14.0.0",
     "bignumber.js": "^9.0.1",
     "bn.js": "^5.2.0",
     "eth-sig-util": "^3.0.1",
