+++ packages/web-client/package.json
@@ -72,7 +72,6 @@
     "@types/rsvp": "^4.0.3",
     "@types/sinon": "^9.0.11",
     "@types/ua-parser-js": "^0.7.36",
-    "@types/web3-provider-engine": "^14.0.1",
     "babel-eslint": "^10.1.0",
     "bn.js": "^5.2.0",
     "broccoli-asset-rev": "^3.0.0",
