+++ packages/web-client/tsconfig.json
@@ -33,6 +33,7 @@
       ],
       "@ember/destroyable": ["../../node_modules/ember-destroyable-polyfill"],
       "@cardstack/web-client/mirage/*": ["mirage/*"],
+      "web3-provider-engine": ["../../types/web3-provider-engine"],
     }
   },
   "include": [
