--- packages/web-client/types/web3-provider-engine/subproviders.d.ts
@@ -1,29 +1,0 @@
-declare module 'web3-provider-engine/subproviders/cache' {
-  import CacheSubprovider from 'web3-provider-engine/subproviders/cache';
-  export default CacheSubprovider;
-}
-
-declare module 'web3-provider-engine/subproviders/fixture' {
-  import FixtureSubprovider from 'web3-provider-engine/subproviders/fixture';
-  export default FixtureSubprovider;
-}
-
-declare module 'web3-provider-engine/subproviders/filters' {
-  import FilterSubprovider from 'web3-provider-engine/subproviders/filters';
-  export default FilterSubprovider;
-}
-
-declare module 'web3-provider-engine/subproviders/hooked-wallet' {
-  import HookedWalletSubprovider from 'web3-provider-engine/subproviders/hooked-wallet';
-  export default HookedWalletSubprovider;
-}
-
-declare module 'web3-provider-engine/subproviders/nonce-tracker' {
-  import NonceSubprovider from 'web3-provider-engine/subproviders/nonce-tracker';
-  export default NonceSubprovider;
-}
-
-declare module 'web3-provider-engine/subproviders/subscriptions' {
-  import SubscriptionsSubprovider from 'web3-provider-engine/subproviders/subscriptions';
-  export default SubscriptionsSubprovider;
-}
